+++ utils/image/include/image/image.hpp
@@ -105,7 +105,7 @@ private:
 typedef ImageBMP<uint8_t> ImageBMP8Bit;
 typedef ImageBMP<uint32_t> ImageBMP32Bit;
 
-template <typename T> int size_in_bytes(const Image<T> &i) {
+template <typename T> size_t size_in_bytes(const Image<T> &i) {
   return i.size_in_bytes();
 }
